Minor type changes

+++ DDMathParser/DDMathStringToken.m
@@ -65,7 +65,7 @@
         [d appendFormat:@"arity:%@, ", arityNames[arity]];
         
         NSInteger precedence = [self operatorPrecedence];
-        [d appendFormat:@"precedence:%ld, ", precedence];
+        [d appendFormat:@"precedence:%ld, ", (long)precedence];
         
         DDOperatorAssociativity assoc = [self operatorAssociativity];
         NSString *assocNames[2] = { @"LEFT", @"RIGHT" };
